feat: refactor deposit handling by introducing Deposit type and updating parsing logic
