+++ utils/arguments.ts
@@ -1,6 +1,11 @@
 import { program } from 'commander';
-import { Permit, RoleAssignment, Tier } from 'types';
+import { Permit, RoleAssignment, Tier, Deposit } from 'types';
 import { parseEther } from 'viem';
+
+import { toHex } from './proof/merkle-utils.js';
+
+const toCamelCase = (str: string): string =>
+  str.replace(/_([a-z])/g, (_, char) => char.toUpperCase());
 
 export const stringToBigIntArray = (value: string) => {
   return value.split(',').map(BigInt);
@@ -39,4 +44,32 @@ export const parseTiers = (value: string) => {
 
 export const parseTier = (value: string) => {
   return JSON.parse(value) as Tier;
+};
+
+export const parseDepositArray = (str: string): Deposit[] => {
+  const trimmed = str.trim();
+  if (!trimmed) {
+    return [];
+  }
+
+  const parsed = JSON.parse('[' + trimmed + ']', (key, value) => {
+    if (key === '') return value; // root array
+    if (key === 'amount') return BigInt(value) * BigInt(10 ** 9); // to wei
+    if (typeof value === 'string') {
+      return toHex(value);
+    }
+    return value;
+  });
+
+  // Convert keys to camelCase
+  const camelCased: Deposit[] = parsed.map((obj: any) => {
+    const newObj: any = {};
+    for (const key in obj) {
+      const camelKey = toCamelCase(key);
+      newObj[camelKey] = obj[key];
+    }
+    return newObj;
+  });
+
+  return camelCased;
 };
