+++ utils/index.ts
@@ -9,7 +9,6 @@ export * from './prompts/index.js';
 export * from './spinner/index.js';
 export * from './read-programs-by-abi.js';
 export * from './get-commands.js';
-export * from './parse-string.js';
 export * from './get-deposit-data-root.js';
 export * from './arguments.js';
 export * from './report.js';
